+++ pkg/security/probe/field_handlers_ebpf.go
@@ -3,6 +3,8 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-present Datadog, Inc.
 
+//go:build linux
+
 // Package probe holds probe related files
 package probe
